Convert sex attribute to IntEnum and update documentation.
The sex attribute of a CURP will now be represented through an IntEnum,
which seems more appropiate as part of the API, and still compatible
with the past approach.

The documentation that depended on the sex attribute has been updated.

Shields in both the README and the documentation index have been
tweaked. Two new shields were added:

- Downloads per month in PyPI
- Python versions the project is compatible with